refactor(store): extract memory impl into separate module
